Integração da API com a tela de tasks, funções de adição de tasks, edição de tasks e exclusão de tasks.

+++ SistemaNoivosApi/src/main/java/br/com/lucas/sistemanoivosapi/Controller/TaskController.java
@@ -3,13 +3,14 @@ package br.com.lucas.sistemanoivosapi.Controller;
 import br.com.lucas.sistemanoivosapi.Model.Task;
 import br.com.lucas.sistemanoivosapi.Repository.TaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("/api/tasks")
-@CrossOrigin(origins = "http://localhost:4200") // Abre permissão para o Angular
+@CrossOrigin(origins = "*")
 public class TaskController {
 
     @Autowired
@@ -26,22 +27,26 @@ public class TaskController {
     }
 
     @PutMapping("/{id}")
-    public Task updateTask(@PathVariable Long id, @RequestBody Task taskDetails) {
-        Task task = taskRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Tarefa não encontrada"));
-
-        task.setTitle(taskDetails.getTitle());
-        task.setDate(taskDetails.getDate());
-        task.setCategory(taskDetails.getCategory());
-        task.setCategoryTheme(taskDetails.getCategoryTheme());
-        task.setCompleted(taskDetails.isCompleted());
-        task.setOverdue(taskDetails.isOverdue());
-
-        return taskRepository.save(task);
+    public ResponseEntity<Task> updateTask(@PathVariable Long id, @RequestBody Task taskDetails) {
+        return taskRepository.findById(id)
+                .map(task -> {
+                    task.setTitle(taskDetails.getTitle());
+                    task.setDate(taskDetails.getDate());
+                    task.setCategory(taskDetails.getCategory());
+                    task.setCategoryTheme(taskDetails.getCategoryTheme());
+                    task.setCompleted(taskDetails.isCompleted());
+                    task.setOverdue(taskDetails.isOverdue());
+                    Task updatedTask = taskRepository.save(task);
+                    return ResponseEntity.ok(updatedTask);
+                }).orElse(ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/{id}")
-    public void deleteTask(@PathVariable Long id) {
-        taskRepository.deleteById(id);
+    public ResponseEntity<Void> deleteTask(@PathVariable Long id) {
+        if (taskRepository.existsById(id)) {
+            taskRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
